templates + more info
Don't forget that putting templates inside the accounts forlder or app will not get read you must put the templates folder under the main app which is crm app not CRM so CRM/crm/templates. Adding functions to account.views.py and adding the path to the urls.py account file.

So when a website user visists the crm site they first see the home page which is accounts.urls because it only has '' and because of that main url sends visitors to the account url and the account.urls.py activates path('', views.home) by going to account.views.py and execute the home function by returning the HttpResponse('home'). For now the site will display 'home'